chore: make the app.go code generic for different execution clients and minor improvements
