+++ execution/evm/abci_handler.go
@@ -1,7 +1,6 @@
 package evm
 
 import (
-	"context"
 	"encoding/json"
 	"fmt"
 	"time"
@@ -14,19 +13,7 @@ import (
 	"github.com/ethereum/go-ethereum/core/types"
 )
 
-type EVMABCIHandler struct {
-	ctx    context.Context
-	client *EVMClient
-}
-
-func NewEVMABCIHandler(ctx context.Context, evmClient *EVMClient) (*EVMABCIHandler, error) {
-	return &EVMABCIHandler{
-		ctx:    ctx,
-		client: evmClient,
-	}, nil
-}
-
-func (h *EVMABCIHandler) PrepareProposal() sdk.PrepareProposalHandler {
+func (c *EVMClient) PrepareProposal() sdk.PrepareProposalHandler {
 	return func(ctx sdk.Context, req *abci.RequestPrepareProposal) (*abci.ResponsePrepareProposal, error) {
 		// todo: put retries for rpc and engine api calls
 
@@ -37,14 +24,14 @@ func (h *EVMABCIHandler) PrepareProposal() sdk.PrepareProposalHandler {
 		json.Unmarshal(bz, &params)
 		fmt.Printf("PARAMS IN PREPARE PROPOSAL: %+v\n", params)
 
-		lbn, err := h.client.GetLatestBlockNumber()
+		lbn, err := c.GetLatestBlockNumber()
 		if err != nil {
 			return nil, err
 		}
 
 		fmt.Printf("LATEST BLOCK NUMBER: %v\n", lbn.Int64())
 
-		bh, err := h.client.GetBlockByNumber(lbn, false)
+		bh, err := c.GetBlockByNumber(lbn, false)
 		if err != nil {
 			return nil, err
 		}
@@ -76,14 +63,14 @@ func (h *EVMABCIHandler) PrepareProposal() sdk.PrepareProposalHandler {
 			Withdrawals:           []*types.Withdrawal{},
 		}
 
-		fcres, err := h.client.ForkchoiceUpdated(state, attr)
+		fcres, err := c.ForkchoiceUpdated(state, attr)
 		if err != nil {
 			return nil, err
 		}
 
 		fmt.Printf("ForkchoiceUpdated response: %+v\n", fcres)
 
-		payloadExData, err := h.client.GetPayload(*fcres.PayloadID)
+		payloadExData, err := c.GetPayload(*fcres.PayloadID)
 		if err != nil {
 			return nil, err
 		}
@@ -106,7 +93,7 @@ func (h *EVMABCIHandler) PrepareProposal() sdk.PrepareProposalHandler {
 	}
 }
 
-func (h *EVMABCIHandler) ProcessProposal() sdk.ProcessProposalHandler {
+func (c *EVMClient) ProcessProposal() sdk.ProcessProposalHandler {
 	return func(ctx sdk.Context, req *abci.RequestProcessProposal) (*abci.ResponseProcessProposal, error) {
 		// Once you receive the prepare proposal response make a new payload request to the EVM.
 		var (
@@ -125,7 +112,7 @@ func (h *EVMABCIHandler) ProcessProposal() sdk.ProcessProposalHandler {
 		fmt.Printf("RECIEVED EXECUTION PAYLOAD: %+v\n", executableData)
 		fmt.Printf("RECIEVED PAYLOAD ATTRIBUTES: %+v\n", attributes)
 
-		res, err := h.client.NewPayload(*executableData, []common.Hash{}, (common.Hash)(attributes.ParentBeaconBlockRoot))
+		res, err := c.NewPayload(*executableData, []common.Hash{}, (common.Hash)(attributes.ParentBeaconBlockRoot))
 		if err != nil {
 			return nil, err
 		}
@@ -138,7 +125,7 @@ func (h *EVMABCIHandler) ProcessProposal() sdk.ProcessProposalHandler {
 	}
 }
 
-func (h *EVMABCIHandler) PreBlock() sdk.PreBlocker {
+func (c *EVMClient) PreBlock() sdk.PreBlocker {
 	return func(ctx sdk.Context, req *abci.RequestFinalizeBlock) (*sdk.ResponsePreBlock, error) {
 		executableData := &ExecutableData{}
 
@@ -151,7 +138,7 @@ func (h *EVMABCIHandler) PreBlock() sdk.PreBlocker {
 			FinalizedBlockHash: executableData.ParentHash,
 		}
 
-		fcres, err := h.client.ForkchoiceUpdated(state, nil)
+		fcres, err := c.ForkchoiceUpdated(state, nil)
 		if err != nil {
 			return nil, err
 		}
@@ -164,7 +151,7 @@ func (h *EVMABCIHandler) PreBlock() sdk.PreBlocker {
 	}
 }
 
-func (h *EVMABCIHandler) EndBlock() sdk.EndBlocker {
+func (c *EVMClient) EndBlock() sdk.EndBlocker {
 	return func(ctx sdk.Context) (sdk.EndBlock, error) {
 		return sdk.EndBlock{}, nil
 	}
